Initialize entity service with entity creation method
This commit introduces the EntityService as well as a method for
generating NPCs given a standard EntityDefinition. The EntityDefintion
is essentially a wrapper around primitives that map back to the domain
objects that make up an NPC.

In the future, I will need to take into account future needs such as
generating Heros for hire and handling various other bits.